Add Stakes & Wallet codefile, Fix typos, refactor POS related func.
POS related codefiles are now connected to the cmake build system.

Typos and wrong function implementations are now fixed,
so that they does not throw compile error.
AsyncEventDataManager should be refactored
because there's so many functionalities specific to each modules.
For now, SetMinedPOSBlock and GetMinedPOSBlock is added for easy implementation.
Some kind of polymorphism for data transferring should be implemented in the future.

Wallet codefile is added for key-value management.
For now, it will be planned to support PoS-stakes-related features,
but that should be expanded to the Transaction sender and receiver usage in the future.

Stakes codefile is added for PoS leader selection supports.
It is empty code for now and it will be implemented soon.